+++ contrib/Config/BladeAndBlunt.ini
@@ -33,6 +33,7 @@ sModFileName = BladeAndBlunt.esp
 ;Injected Forms
 
 BashStaminaPerkFormId = 0xADA510
+BashStaminaPerk25FormId = 0xADA513
 BlockStaminaPerkFormId = 0xADA509
 BlockStaggerPerkFormId = 0xADA511
 
@@ -41,13 +42,3 @@ DualWieldReplacementKeyword = 0xADA172
 InjuryChance25GlobalFormId= 0xADA180
 InjuryChance50GlobalFormId= 0xADA181
 InjuryChance90GlobalFormId= 0xADA182
-
-;Injury Settings
-
-;Only relevant for Starfrost integration
-
-fInjury1AVPercent=0.10
-fInjury2AVPercent=0.25
-fInjury3AVPercent=0.50
-
-fInjuryUpdateFrequency=0.50
