+++ contrib/PluginRelease1130/skse/plugins/BladeAndBlunt.ini
@@ -1,0 +1,44 @@
+bEnableInjuries = true
+bEnableInjuriesOnlyWithSM = false
+bZeroAllWeaponStagger = true
+bArmorRatingScalingEnabled = true
+bReplaceNewAttackTypeKeywords = true
+bLevelBasedDifficulty = false
+
+;Don't edit anything below this comment unless you know what you're doing! 
+
+IsAttackingSpellFormId = 0x000821
+IsBlockingSpellFormId = 0x000825
+IsCastingSpellFormId = 0x00089F
+
+BowStaminaSpellFormId = 0x000820
+XbowStaminaSpellFormId = 0x000875
+
+PowerAttackStopSpellFormId = 0x0008D5
+
+IsSneakingSpellFormId = 0x000823
+IsSprintingSpellFormId = 0x0008A7
+
+MountSprintingSpellFormId = 0x0008AD
+
+JumpSpellFormID = 0x0008BF
+SwimmingSpellFormId = 0x008CB
+
+InjurySpell1FormID = 0x00084A
+InjurySpell2FormID = 0x00084B
+InjurySpell3FormID = 0x00084D
+
+sModFileName = BladeAndBlunt.esp
+
+;Injected Forms
+
+BashStaminaPerkFormId = 0xADA510
+BashStaminaPerk25FormId = 
+BlockStaminaPerkFormId = 0xADA509
+BlockStaggerPerkFormId = 0xADA511
+
+DualWieldReplacementKeyword = 0xADA172
+
+InjuryChance25GlobalFormId= 0xADA180
+InjuryChance50GlobalFormId= 0xADA181
+InjuryChance90GlobalFormId= 0xADA182
