+++ include/Events.h
@@ -1,6 +1,6 @@
 #pragma once
-#include <Conditions.h>
-#include <InjuryApplicationManager.h>
+#include <Utility.h>
+#include <injury/InjuryApplicationManager.h>
 #include <RecentHitEventData.h>
 
 class OnHitEventHandler : public RE::BSTEventSink<RE::TESHitEvent>
@@ -42,7 +42,7 @@ public:
 				//Something is effed with power attacks. The source isnt coming through and casting as a weapon and the hit flags are empty
 				//We can work around it like this
 				bool powerAttackMelee = false;
-				if (a_event->flags.any(RE::TESHitEvent::Flag::kPowerAttack) || Conditions::IsPowerAttacking(causeActor)) {
+                if (a_event->flags.any(RE::TESHitEvent::Flag::kPowerAttack) || Utility::IsPowerAttacking(causeActor)) {
 					bool rightIsMeleeWeapon = false;
 					if (auto rightHandForm = causeActor->GetEquippedObject(false)) {
 						if (rightHandForm->IsWeapon() && rightHandForm->As<RE::TESObjectWEAP>()->IsMelee()) {
@@ -75,19 +75,16 @@ public:
                     }
 				}
 
-				auto leftHand = targetActor->GetEquippedObject(true);
-
 				bool blockedMeleeHit = false;
 				if (!a_event->projectile && 
 					((attackingWeapon && attackingWeapon->IsMelee()) || powerAttackMelee) &&
 					isBlocking) {
 					blockedMeleeHit = true;
 				}
-				
-				//Shield Stagger
-				if (leftHand && leftHand->IsArmor() && blockedMeleeHit){
+
+                if (blockedMeleeHit) {
+				    //Stagger
 					ProcessHitEventForBlockStagger(targetActor, causeActor);
-				} else if (blockedMeleeHit) {
 					//Parry
 					ProcessHitEventForParry(targetActor,causeActor);
 				}
@@ -100,8 +97,8 @@ public:
 	static void ProcessHitEventForParry(RE::Actor* target, RE::Actor* aggressor)
 	{
 		auto settings = Settings::GetSingleton();
-		if (Conditions::PlayerHasActiveMagicEffect(settings->MAG_ParryWindowEffect)) {
-			Conditions::ApplySpell(target, aggressor, settings->MAGParryStaggerSpell);
+        if (Utility::PlayerHasActiveMagicEffect(settings->MAG_ParryWindowEffect)) {
+            Utility::ApplySpell(target, aggressor, settings->MAGParryStaggerSpell);
 		}
 	}
 
@@ -109,10 +106,10 @@ public:
 	{
 		auto settings = Settings::GetSingleton();
 		if (target->HasPerk(settings->BlockStaggerPerk)) {
-			Conditions::ApplySpell(target, aggressor, settings->MAGBlockStaggerSpell2);
+            Utility::ApplySpell(target, aggressor, settings->MAGBlockStaggerSpell2);
 		}
 		else {
-			Conditions::ApplySpell(target, aggressor, settings->MAGBlockStaggerSpell);
+			Utility::ApplySpell(target, aggressor, settings->MAGBlockStaggerSpell);
 		}
 	}
 
@@ -144,94 +141,6 @@ public:
 	}
 };
 
-class AnimationGraphEventHandler :
-    public RE::BSTEventSink<RE::BSAnimationGraphEvent>,
-    public RE::BSTEventSink<RE::TESObjectLoadedEvent>,
-    public RE::BSTEventSink<RE::TESSwitchRaceCompleteEvent>
-
-{
-public:
-
-    static AnimationGraphEventHandler* GetSingleton()
-    {
-        static AnimationGraphEventHandler singleton;
-        return &singleton;
-    }
-
-    const char* jumpAnimEventString = "JumpUp";
-
-    //Anims
-    RE::BSEventNotifyControl ProcessEvent(const RE::BSAnimationGraphEvent* a_event, [[maybe_unused]] RE::BSTEventSource<RE::BSAnimationGraphEvent>* a_eventSource) override
-    {
-        if (!a_event) {
-            return RE::BSEventNotifyControl::kContinue;
-        }
-
-        if (!a_event->tag.empty() && a_event->holder && a_event->holder->As<RE::Actor>()) {
-            if (std::strcmp(a_event->tag.c_str(), jumpAnimEventString) == 0) {
-
-                HandleJumpAnim();
-            }
-        }
-        
-        return RE::BSEventNotifyControl::kContinue;
-    }
-
-    void HandleJumpAnim()
-    {
-        auto settings = Settings::GetSingleton();
-        auto player   = RE::PlayerCharacter::GetSingleton();
-        if (!player->IsGodMode()) {
-            Conditions::ApplySpell(player, player, settings->jumpSpell);
-        }
-    }
-
-    //Object load
-    RE::BSEventNotifyControl ProcessEvent(const RE::TESObjectLoadedEvent* a_event, [[maybe_unused]] RE::BSTEventSource<RE::TESObjectLoadedEvent>* a_eventSource) override
-    {
-        if (!a_event) {
-            return RE::BSEventNotifyControl::kContinue;
-        }
-        
-        const auto actor = RE::TESForm::LookupByID<RE::Actor>(a_event->formID);
-        if (!actor || !actor->IsPlayerRef()) {
-            return RE::BSEventNotifyControl::kContinue;
-        }
-
-        //Register for anim event
-        actor->AddAnimationGraphEventSink(AnimationGraphEventHandler::GetSingleton());
-
-        return RE::BSEventNotifyControl::kContinue;
-    }
-
-    // Race Switch
-    RE::BSEventNotifyControl ProcessEvent(const RE::TESSwitchRaceCompleteEvent* a_event, [[maybe_unused]] RE::BSTEventSource<RE::TESSwitchRaceCompleteEvent>* a_eventSource) override
-    {
-        if (!a_event) {
-            return RE::BSEventNotifyControl::kContinue;
-        }
-
-        const auto actor = a_event->subject->As<RE::Actor>();
-        if (!actor || !actor->IsPlayerRef()) {
-            return RE::BSEventNotifyControl::kContinue;
-        }
-
-        // Register for anim event
-        actor->AddAnimationGraphEventSink(AnimationGraphEventHandler::GetSingleton());
-
-        return RE::BSEventNotifyControl::kContinue;
-    }
-
-    static void Register()
-    {
-
-        //Register for load event, then in the load event register for anims
-        RE::ScriptEventSourceHolder* eventHolder = RE::ScriptEventSourceHolder::GetSingleton();
-        eventHolder->AddEventSink<RE::TESObjectLoadedEvent>(GetSingleton());
-        eventHolder->AddEventSink<RE::TESSwitchRaceCompleteEvent>(GetSingleton());
-    }
-};
-
 class WeaponFireHandler
 {
 public:
@@ -254,9 +163,16 @@ public:
 
 		auto source = a_source->As<RE::Actor>();
 
-		if (source->IsPlayerRef() && a_weapon->IsCrossbow()) {
-			Conditions::ApplySpell(source, source,Settings::GetSingleton()->MAGCrossbowStaminaDrainSpell);
+		if (source->IsPlayerRef() && (a_weapon->IsCrossbow() || a_weapon->IsBow())) {
+			Utility::ApplySpell(source, source,Settings::GetSingleton()->MAGCrossbowStaminaDrainSpell);
+            RE::BSTSmartPointer<RE::BSAnimationGraphManager> manager;
+
+            auto playerCamera = RE::PlayerCamera::GetSingleton();
+            if (playerCamera->bowZoomedIn) {
+                source->NotifyAnimationGraph("attackStop");
+            }
 		}
+
 	}
 	inline static REL::Relocation<decltype(WeaponFire)> _Weapon_Fire;
 };
