+++ src/Settings.cpp
@@ -23,10 +23,6 @@ void Settings::LoadSettings()
 	armorScalingEnabled = ini.GetBoolValue("", "bArmorRatingScalingEnabled", true);
 	replaceAttackTypeKeywords = ini.GetBoolValue("", "bReplaceNewAttackTypeKeywords", true);
 
-	injury1AVPercent = static_cast<float>(ini.GetDoubleValue("", "fInjury1AVPercent", 0.10));
-	injury2AVPercent = static_cast<float>(ini.GetDoubleValue("", "fInjury2AVPercent", 0.25));
-	injury3AVPercent = static_cast<float>(ini.GetDoubleValue("", "fInjury3AVPercent", 0.50));
-
 	std::string fileName(ini.GetValue("", "sModFileName", ""));
 	FileName = fileName;
 }
@@ -172,6 +168,9 @@ void Settings::LoadForms()
     std::string bashPerkFormId((ini.GetValue("", "BashStaminaPerkFormId", "")));
     BashStaminaPerk = LoadFormPointerFromIni<RE::BGSPerk>(bashPerkFormId, "Update.esm");
 
+    std::string bashPerk25FormId((ini.GetValue("", "BashStaminaPerk25FormId", "")));
+    BashStaminaPerk25 = LoadFormPointerFromIni<RE::BGSPerk>(bashPerk25FormId, "Update.esm");
+    
     std::string blockPerkFormId((ini.GetValue("", "BlockStaminaPerkFormId", "")));
     BlockStaminaPerk = LoadFormPointerFromIni<RE::BGSPerk>(blockPerkFormId, "Update.esm");
 
@@ -183,12 +182,23 @@ void Settings::LoadForms()
 
 	//Hardcoded loads
 	MAGParryControllerSpell = dataHandler->LookupForm(ParseFormID("0x817"), FileName)->As<RE::SpellItem>();
+
 	MAGParryStaggerSpell = dataHandler->LookupForm(ParseFormID("0x816"), FileName)->As<RE::SpellItem>();
 	MAGBlockStaggerSpell = dataHandler->LookupForm(ParseFormID("0x855"), FileName)->As<RE::SpellItem>();
 	MAGBlockStaggerSpell2 = dataHandler->LookupForm(ParseFormID("0x858"), FileName)->As<RE::SpellItem>();
+
+    MAG_ChargedShotSpell01 = dataHandler->LookupForm(ParseFormID("0x1B"), FileName)->As<RE::SpellItem>();
+    MAG_ChargedShotSpell02 = dataHandler->LookupForm(ParseFormID("0x1E"), FileName)->As<RE::SpellItem>();
+    MAG_ChargedShotSpell03 = dataHandler->LookupForm(ParseFormID("0x1F"), FileName)->As<RE::SpellItem>();
+
+    MAG_ChargedShotTimer01 = dataHandler->LookupForm(ParseFormID("0x28"), FileName)->As<RE::TESGlobal>();
+    MAG_ChargedShotTimer02 = dataHandler->LookupForm(ParseFormID("0x2C"), FileName)->As<RE::TESGlobal>();
+
 	MAGCrossbowStaminaDrainSpell = dataHandler->LookupForm(ParseFormID("0x873"), FileName)->As<RE::SpellItem>();
+
 	MAG_InjuryCooldown1 = dataHandler->LookupForm(ParseFormID("0x84F"), FileName)->As<RE::EffectSetting>();
 	MAG_InjuryCooldown2 = dataHandler->LookupForm(ParseFormID("0x850"), FileName)->As<RE::EffectSetting>();
+
 	MAG_ParryWindowEffect = dataHandler->LookupForm(ParseFormID("0x815"), FileName)->As<RE::EffectSetting>();
 	MAG_InjuriesSMOnly = dataHandler->LookupForm(ParseFormID("0x88E"), FileName)->As<RE::TESGlobal>();
 
@@ -196,6 +206,8 @@ void Settings::LoadForms()
 	MAG_PowerAttackReplacement = dataHandler->LookupForm(ParseFormID("0xA9B"), FileName)->As<RE::TESGlobal>();
 	MAG_InjuryAndRest = dataHandler->LookupForm(ParseFormID("0x83F"), FileName)->As<RE::TESGlobal>();
 	HealthPenaltyUIGlobal = dataHandler->LookupForm(RE::FormID(0x2EDE), "Update.esm")->As<RE::TESGlobal>();
+
+    MAG_DifficultyGlobal = dataHandler->LookupForm(RE::FormID(0xADA617), "Update.esm")->As<RE::TESGlobal>();
 
 	auto smGlobal = dataHandler->LookupForm(RE::FormID(0x826), "ccqdrsse001-survivalmode.esl");
